refactor: cleaning up of config data types (incomplete)
