+++ network/blocker.py
@@ -8,6 +8,7 @@ from typing import Optional
 import pydivert
 
 from network import sessioninfo
+from network.MinimalPacket import safe_pickle_packet
 from util.DynamicBlacklist import ip_in_cidr_block_set
 
 debug_logger = logging.getLogger("debugger")
@@ -137,7 +138,7 @@ class AbstractPacketFilter(ABC):
 
                     if self.session_info is not None:
                         self.session_info.add_packet(
-                            sessioninfo.safe_pickle_packet(packet), allowed=decision
+                            safe_pickle_packet(packet), allowed=decision
                         )
 
                     if self.debug_print_decisions:
